+++ src/CommPlats/Loom_nRF.h
@@ -51,7 +51,7 @@ public:
 	Loom_nRF(
 			const char*		module_name			= "nRF",
 			uint16_t		max_message_len		= 120,
-			uint16_t		address 			= 0,
+			uint8_t			address 			= 0,
 			uint8_t			data_rate			= 1,
 			uint8_t			power_level			= 0,
 			uint8_t			retry_count 		= 3,
@@ -72,13 +72,13 @@ public:
 /*@{*/ //======================================================================
 
 	bool		receive(JsonObject json) override;
-	bool		send(JsonObject json, uint16_t destination) override;
+	bool		send(JsonObject json, uint8_t destination) override;
 	void		broadcast(JsonObject json) override;
 	void		add_config(JsonObject json) override;
 
 	// manually expose superclass version of log() that gets json from
 	// linked LoomManager, calling this classes implementation of 
-	// 'send(JsonObject)' and 'send(JsonObject, uint16_t)', which is pure virtual in superclass
+	// 'send(JsonObject)' and 'send(JsonObject, uint8_t)', which is pure virtual in superclass
 	using LoomCommPlat::send; 
 	using LoomCommPlat::receive; 
 	using LoomCommPlat::broadcast; 
@@ -93,7 +93,7 @@ public:
 ///@name	GETTERS
 /*@{*/ //======================================================================
 
-	uint		get_address() override;
+	uint8_t		get_address() override;
 
 	/// Get multicast(broadcast) level.
 	/// nRF has varying degrees of broadcast corresponding to 
@@ -105,7 +105,7 @@ public:
 ///@name	SETTERS
 /*@{*/ //======================================================================
 
-	void		set_address(uint addr) override;
+	void		set_address(uint8_t addr) override;
 
 	/// Set multicast(broadcast) level.
 	/// nRF has varying degrees of broadcast corresponding to
